fix: fix compilation error in arm platform
